Disable relative line numbers again
Scrolling through a file or adding many new lines when using
`relativenumber` can easily result in a single CPU core being used for
about 80%. On a desktop this is not that big of a deal, but on my laptop
it can easily result in the battery being drained faster, when done
often enough.

Using incremental searching, the use of relative numbers is not that
necessary any more. I also found that relative numbers stop being useful
once you have to jump 20 or so lines, as at that point I'd spend quite a
bit of time scanning what number of lines to jump.