Replace Deoplete with CoC + ALE
CoC is used for language server integration and completion. ALE is used
for linting and formatting, as it supports a wider variety of file
types. CoC diagnostics are displayed using ALE, keeping the output
consistent between both tools.

For Python I use coc-jedi, as the coc-python extension is missing too
many features (e.g. renaming doesn't work), and uses quite a bit of
memory in the main Node process.