Use nvim-autopairs and a custom linting setup
nvim-autopairs has better support for adding spaces in pairs, and keeps
track of pair balancing when exiting insert mode.

The custom linter setup is based on nvim-lint, but with some added
functionality and changes such as:

- Better error messages
- A more flexible way of specifying the commands and arguments
- No support for STDIN, because I don't need it
- Being able to conditionally enable linters (e.g. based on the presence
  of a config file)
- Automatically running RuboCop with `bundle exec`, but only if it's in
  the Gemfile

Most important of all, since it's a custom solution I have full control
over how it works.