Show the LSP status in the tabline
Since this status is global instead of tied to a buffer, showing it in a
statusline doesn't make sense. This would also result in every
statusline showing the same info.

Instead, the LSP status is shown in the tabline. As tablines aren't
redrawn as frequently, we need an extra autocmd to refresh it when new
progress messages are available.