Remove Bubblewrap code
This was a fun experiment, and I mostly tracked it in Git in case I want
to refer back to it in the future, but it's too annoying to use in
practise. In addition, the security benefits are debatable, as both
NeoVim and its plugins have unrestricted editor and file system access
anyway. But at least I now know how Bubblewrap works.