Use a vimscript expression for statuscolumn again
When using a Lua function, it appears options such as
`vim.wo.relativenumber` don't return the same value as
`&relativenumber`. The result is that opening e.g. Telescope (which
disables relativenumber) also disables line numbers of the previous
window. This isn't the case when using &relativenumber in a Vimscript
expression.