Stop signing Git commits
After doing so for years, I've concluded there's no value in doing this.

First, most tooling out there (including Git's own CLI) don't display
the signing status. As a result, most likely aren't aware of you signing
commits or that it's even possible to do so.

Second, it doesn't integrate well with conventional workflows:
ammending, rebasing, squashing, etc, all rewrite commits. These steps
overwrite signatures, based on the person performing them. Worse, if
done using GitHub, a global key shared across everybody is used to sign
the commits.

This leads me to the following problem: signing is useless if
distributing the keys is a pain in the ass. Git doesn't do anything to
make this easier, resulting in ad-hoc solutions to distribute public
keys ("just go to this website, then run a bunch of commands to add
it"). 99% of the people can't be bothered with this, and rightfully so.
In case of GitHub's use of a global GPG key for certain operations,
trusting the key also means trusting _every_ commit signed with it, even
if you don't know the author.

Finally, the whole process requires you to maintain the signing keys and
share these across all your environments you may create commits from.
This is a little easier using SSH keys, but it's still extra work to do
and to keep in mind.

I personally used signing as a way of saying "I truly authored this
commit" in case somebody tries to "impersonate" me on e.g. GitHub. In
reality this doesn't really happen, and there are plenty of other ways
one could prove they are or aren't the author of a commit (i.e. by
showing they never had access to the repository).

In other words: signing commits is basically a cargo-cult behavior one
falls prey to because "some blog post said I should be using it", not
because it actually solves problems.