Disable rust-analyzer live diagnostics
Enabling this results in lots of duplicate diagnostics upon saving, as
both rust-analyzer and cargo check produce the same kind (but slightly
differently formatted) diagnostics. In addition, the diagnostics from
rust-analyzer tend to use weird source locations (e.g. the end of a
method call for argument errors, instead of the start).