Remove ccc.nvim
A color picker is already a bit of a niche, but this plugin is _very_
chatty when it comes to its language server integration, triggering an
avalanche of LSP calls even when this is completely redundant.

Since a text based color picker is a bit clunky to use in the first
place, let's just get rid of this plugin.