Replace oil with mini.files
Moving files around oil buffers is a bit clunky the moment you need to
move it to a different directory, as this requires opening multiple
windows. In addition, remembering where you are in the hierarchy is a
bit confusing.

Since I already have mini.nvim installed I might as well give mini.files
a try for a while.